pyproject.toml
[tool.poetry]
name = "shushukun-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.1"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
SQLAlchemy = "^1.4.29"
aiomysql = "^0.0.22"

[tool.poetry.dev-dependencies]
pytest-asyncio = "^0.16.0"
aiosqlite = "^0.17.0"
httpx = "^0.21.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
